Add password prompt and improve error handling in main.go
main.go was modified to improve security and error handling. A password prompt was added to the main function, and the logic for handling unsuccessful password entry was improved. The script was also adjusted to offer a more engaging interaction with the user. Error handling was centralized to a pauses function to standardize behavior. The .gitignore file was further modified to stop ignoring .exe files. These changes aim to improve usability and robustness of the software.