Update README.md with installation instructions
A clear, step-by-step installation guide has been added to README.md for newcomer ease and understanding. The last section was rewritten to include the introduction of the new feature and its installation process. The aim is to provide thorough information to the users about how to get the software up and running, thus improving the overall user experience.